+++ utils/instructions/PsyFinance/index.ts
@@ -17,6 +17,7 @@ import { OptionMarket, OptionMarketWithKey } from './types'
 
 export { PsyAmericanIdl } from './PsyAmericanIdl'
 export * from './types'
+export * from './hooks'
 
 export const PSY_AMERICAN_PROGRAM_ID = new PublicKey(
   'R2y9ip6mxmWUj4pt54jP2hz2dgvMozy9VTSwMWE7evs'
